search working and delete validation working
User can't delete posts not made by them.

Users can search by either city, restaurant name, or both

+++ controllers/deleteController.js
@@ -1,15 +1,11 @@
-
 // import module `database` from `../models/db.js`
 const db = require('../models/db.js');
-
-// import module `User` from `../models/UserModel.js`
-const User = require('../models/userModel.js');
 
 // import module `User` from `../models/UserModel.js`
 const Review = require('../models/reviewModel.js');
 
 // import module `User` from `../models/UserModel.js`
-const Restaurant = require('../models/restaurantModel.js');
+const User = require('../models/userModel.js');
 
 //for ObjectId(_id), used to get the Object type of the retrieved _id
 const ObjectId = require('mongodb').ObjectID;
@@ -19,40 +15,31 @@ const ObjectId = require('mongodb').ObjectID;
 const deleteController = {
 
     deleteReview: function (req, res){
-        db.deleteOne(Review, {_id:req.params.reviewID});
-        res.redirect('/');
-
-        // db.findOne(User, {isLoggedIn: true}, 'uname _id', function(loggedUserResult){
-
-        //     console.log(loggedUserResult);
-        //     var _id = loggedUserResult._id;
-        //     var uname = loggedUserResult.uname;
-        //     // console.log(_id);
         
-        //     if(loggedUserResult == !null){
-        //         console.log("TRUEEEE");
-        //         db.findOne(Review, {_id: req.params.reviewID }, '_id restaurantID authorID', function(result){
-        //             var authorID = result.authorID;
 
-        //             console.log(result);
+        db.findOne(User, {isLoggedIn: true}, 'uname _id', function(loggedUserResult){
+            console.log("enter user");
 
-        //             console.log(_id);
-        //             console.log(authorID);
-        //             if(_id == authorID){
-        //                 db.deleteOne(Review, {_id:req.params.reviewID});
-        //                 res.redirect('/');
-        //             }else{
-        //                 render('error', {errormsg: "This review is not yours. You can only delete reviews you made."});
-        //             }
-        //         })
-        //     }else{
-        //         res.render('error', {uname:uname, errormsg: "This review may be not yours or you are not logged in."});
-        //     }
-                
+            db.findOne(Review,{_id:req.params.reviewID}, 'authorID', function(reviewResult){
 
-        // })
+                //if the currently logged in user is the author of the review
+                if(ObjectId(loggedUserResult._id).toString() == reviewResult.authorID ){
+
+                    db.deleteOne(Review, {_id:req.params.reviewID});
+
+                    //refreshes the page
+                    res.redirect('back');
+                }
+                //else, say that u can't delete reviews you didn't make
+                else{
+                    res.render('error', {errormsg: "You can't delete reviews published by other users!"});
+                }
+
+            } )
+
+
+        })
     }
-
 }
 
 module.exports = deleteController;
